requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.0.12
discord==1.7.3
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
typing-extensions==4.2.0
yarl==1.7.2
